add some unit tests
